% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_aloha.R
\name{say_aloha}
\alias{say_aloha}
\title{SummR test function say_aloha}
\usage{
say_aloha(name, print_to_console = TRUE)
}
\arguments{
\item{name}{(character) Name of the person to say "have a good summer" to}

\item{print_to_console}{(logical) Option to print your message. Defaults to TRUE}
}
\value{
(character) good summer message
}
\description{
SummR test function say_aloha
}
\examples{
#say hello to a friend
say_aloha("Tom")
}
