% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_aloha.R
\name{say_good_summer}
\alias{say_good_summer}
\title{Say Good summer}
\usage{
say_good_summer(name, print_to_console = TRUE)
}
\arguments{
\item{name}{(character) Name of the person to say good summer to}

\item{print_to_console}{(logical) Option to print your message. Defaults to TRUE}
}
\value{
(character) good summer message
}
\description{
Say Good summer
}
\examples{
# say good summer to Eliott
say_aloha("Eliott")

}
