% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayaloha.R
\name{say_aloha}
\alias{say_aloha}
\title{Say Aloha}
\usage{
say_aloha(name, print = TRUE)
}
\arguments{
\item{name}{(character) Name of the person to say good summer to}

\item{print}{(logical) Option to print your message. Defaults to TRUE.p}
}
\value{
(character) good summer message
}
\description{
Say Aloha
}
\examples{
# say good summer to Cristina
say_aloha("Cristina")

}
